refactor: clean split CU and TR summary in SPG show (FLEX-1146)
