+++ frontend/src/components/ui/index.ts
@@ -17,6 +17,7 @@ import {
   Checkbox,
   Content,
   Container,
+  Divider,
   FlexDiv,
   Header,
   FormItem,
@@ -67,6 +68,7 @@ export {
   Container,
   Datepicker,
   DateTimePicker,
+  Divider,
   Dropdown,
   FlexDiv,
   Header,
